@@ -21,45 +21,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.6",
-    "@angular/cdk": "~7.2.2",
-    "@angular/common": "~7.2.6",
-    "@angular/compiler": "~7.2.6",
-    "@angular/core": "~7.2.6",
-    "@angular/elements": "^7.2.6",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "~7.2.6",
-    "@angular/http": "^7.2.6",
-    "@angular/material": "^7.2.2",
-    "@angular/platform-browser": "~7.2.6",
-    "@angular/platform-browser-dynamic": "~7.2.6",
-    "@angular/router": "~7.2.6",
+    "@angular/animations": "~8.1.0",
+    "@angular/cdk": "~8.0.2",
+    "@angular/common": "~8.1.0",
+    "@angular/compiler": "~8.1.0",
+    "@angular/core": "~8.1.0",
+    "@angular/elements": "^8.1.0",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "~8.1.0",
+    "@angular/material": "^8.0.2",
+    "@angular/platform-browser": "~8.1.0",
+    "@angular/platform-browser-dynamic": "~8.1.0",
+    "@angular/router": "~8.1.0",
     "@tibco/spotfire-wrapper": "^0.2.6",
-    "angular6-json-schema-form": "^7.0.5",
+    "angular6-json-schema-form": "^7.3.0",
     "chart.js": "^2.8.0",
     "chartjs-plugin-datalabels": "^0.6.0",
     "chartjs-plugin-doughnutlabel": "^2.0.3",
-    "core-js": "^2.6.3",
+    "core-js": "^3.1.4",
     "document-register-element": "^1.7.2",
     "hammerjs": "^2.0.8",
-    "ng2-charts": "^2.2.2",
+    "lodash-es": "^4.17.11",
+    "ng2-charts": "^2.3.0",
     "ngx-color-picker": "^7.5.0",
-    "npm": "^6.9.0",
+    "npm": "^6.10.0",
     "papaparse": "^4.6.3",
-    "rxjs": "~6.4.0",
+    "rxjs": "~6.5.2",
     "tslib": "^1.9.3",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.0",
-    "@angular-devkit/build-ng-packagr": "~0.12.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "^7.2.6",
-    "@angular/language-service": "~7.2.6",
+    "@angular-devkit/build-angular": "~0.801.0",
+    "@angular-devkit/build-ng-packagr": "~0.801.0",
+    "@angular/cli": "~8.1.0",
+    "@angular/compiler-cli": "^8.1.0",
+    "@angular/language-service": "~8.1.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.10.39",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.0.0",
@@ -67,13 +67,13 @@
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.7.1",
+    "ng-packagr": "^5.3.0",
     "ng2-charts-schematics": "^0.1.7",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.1",
-    "tsickle": ">=0.34.0",
+    "tsickle": "^0.35.0",
     "tslib": "^1.9.3",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.4"
+    "typescript": "~3.4.5"
   }
 }
