@@ -1,11 +1,11 @@
 {
   "name": "@tibco-tcstk/tc-core-lib",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "scripts": {
     "compodoc": "npx compodoc -p tsconfig.lib.json"
   },
   "peerDependencies": {
-    "@angular/common": "^7.2.6",
-    "@angular/core": "^7.2.6"
+    "@angular/common": "^8.1.0",
+    "@angular/core": "^8.1.0"
   }
 }
