@@ -1,11 +1,12 @@
 {
   "name": "@tibco-tcstk/tc-forms-lib",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "peerDependencies": {
-    "@angular/common": "^7.2.6",
-    "@angular/core": "^7.2.6",
-    "@angular/material": "^7.2.2",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "angular6-json-schema-form": "^7.0.5"
+    "@angular/common": "^8.1.0",
+    "@angular/core": "^8.1.0",
+    "@angular/material": "^8.0.2",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "angular6-json-schema-form": "^7.3.0",
+    "lodash-es": "^4.17.11"
   }
 }
