@@ -1,19 +1,19 @@
 {
   "name": "@tibco-tcstk/tc-liveapps-lib",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "scripts": {
     "compodoc": "npx compodoc -p tsconfig.lib.json"
   },
   "peerDependencies": {
-    "@angular/common": "^7.2.6",
-    "@angular/core": "^7.2.6",
-    "@angular/material": "^7.2.2",
-    "@angular/flex-layout": "^7.0.0-beta.23",
+    "@angular/common": "^8.1.0",
+    "@angular/core": "^8.1.0",
+    "@angular/material": "^8.0.2",
+    "@angular/flex-layout": "^8.0.0-beta.26",
     "ngx-color-picker": "^7.5.0",
     "ng2-charts": "^2.2.2",
     "chartjs-plugin-datalabels": "^0.6.0",
     "chartjs-plugin-doughnutlabel": "^2.0.3",
-    "@tibco-tcstk/tc-core-lib": "^1.0.0",
-    "@tibco-tcstk/tc-forms-lib": "^1.0.0"
+    "@tibco-tcstk/tc-core-lib": "^1.1.0",
+    "@tibco-tcstk/tc-forms-lib": "^1.1.0"
   }
 }
