@@ -1,13 +1,13 @@
 {
   "name": "@tibco-tcstk/tc-spotfire-lib",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "peerDependencies": {
-    "@angular/common": "^7.2.6",
-    "@angular/core": "^7.2.6",
-    "@angular/material": "^7.2.2",
-    "@angular/flex-layout": "^7.0.0-beta.23",
+    "@angular/common": "^8.1.0",
+    "@angular/core": "^8.1.0",
+    "@angular/material": "^8.0.2",
+    "@angular/flex-layout": "^8.0.0-beta.26",
     "ngx-color-picker": "^7.3.0",
-    "@tibco-tcstk/tc-core-lib": "^1.0.0",
+    "@tibco-tcstk/tc-core-lib": "^1.1.0",
     "@tibco/spotfire-wrapper": "^0.2.3"
   }
 }
